# openapi/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Jikū API",
    "description" : "Event invitation, ticketing, RSVP and check-in API.",
    "version" : "v1"
  },
  "servers" : [
    {
      "url" : "http://localhost",
      "description" : "Generated server url"
    }
  ],
  "security" : [
    {
      "bearer-jwt" : [ ]
    }
  ],
  "paths" : {
...
    "/api/v1/rsvp/{token}/transfer" : {
      "post" : {
        "tags" : [
          "rsvp-controller"
        ],
        "operationId" : "transfer",
        "parameters" : [
          {
            "name" : "token",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TransferTicketRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RsvpView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/rsvp/{token}/confirm" : {
      "post" : {
        "tags" : [
          "rsvp-controller"
        ],
        "operationId" : "confirm",
        "parameters" : [
          {
            "name" : "token",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RsvpView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/events/{eventId}/payments/manual" : {
      "get" : {
        "tags" : [
          "manual-payment-controller"
        ],
        "operationId" : "current",
        "parameters" : [
          {
            "name" : "eventId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ManualPaymentInstructions"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [
          "manual-payment-controller"
        ],
        "operationId" : "request",
        "parameters" : [
          {
            "name" : "eventId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ManualPaymentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ManualPaymentInstructions"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/payments/{id}/reject" : {
      "post" : {
        "tags" : [
          "admin-payment-controller"
        ],
        "operationId" : "reject",
        "parameters" : [
          {
            "name" : "id",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RejectPaymentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminPaymentView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/payments/{id}/confirm" : {
      "post" : {
        "tags" : [
          "admin-payment-controller"
        ],
        "operationId" : "confirm_1",
        "parameters" : [
          {
            "name" : "id",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ConfirmPaymentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminPaymentView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/bookings/{id}/cancel" : {
      "post" : {
        "tags" : [
          "admin-booking-controller"
        ],
        "operationId" : "cancel_1",
        "parameters" : [
          {
            "name" : "id",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminBookingCancellationView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/booking-payments/{id}/verify" : {
      "post" : {
        "tags" : [
          "admin-booking-payment-controller"
        ],
        "operationId" : "verify",
        "parameters" : [
          {
            "name" : "id",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminPaymentDeclarationView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/booking-payments/{id}/reject" : {
      "post" : {
        "tags" : [
          "admin-booking-payment-controller"
        ],
        "operationId" : "reject_1",
        "parameters" : [
          {
            "name" : "id",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RejectDeclarationRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminPaymentDeclarationView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/checkin/{token}" : {
      "get" : {
        "tags" : [
          "validator-check-in-controller"
        ],
        "operationId" : "context",
        "parameters" : [
          {
            "name" : "token",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ValidatorContextResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/payments" : {
      "get" : {
        "tags" : [
          "admin-payment-controller"
        ],
        "operationId" : "list_9",
        "parameters" : [
          {
            "name" : "status",
            "in" : "query",
            "required" : false,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "name" : "provider",
            "in" : "query",
            "required" : false,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "name" : "tenantId",
            "in" : "query",
            "required" : false,
            "schema" : {
              "type" : "string",
              "format" : "uuid"
            }
          },
          {
            "name" : "page",
            "in" : "query",
            "required" : false,
            "schema" : {
              "type" : "integer",
              "format" : "int32",
              "default" : 0
            }
          },
          {
            "name" : "size",
            "in" : "query",
            "required" : false,
            "schema" : {
              "type" : "integer",
              "format" : "int32",
              "default" : 20
            }
          }
        ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/AdminPaymentView"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/bookings" : {
      "get" : {
        "tags" : [
          "admin-booking-controller"
        ],
        "operationId" : "list_10",
        "parameters" : [
          {
            "name" : "status",
            "in" : "query",
            "required" : false,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "name" : "page",
            "in" : "query",
            "required" : false,
            "schema" : {
              "type" : "integer",
              "format" : "int32",
              "default" : 0
            }
          },
          {
            "name" : "size",
            "in" : "query",
            "required" : false,
            "schema" : {
              "type" : "integer",
              "format" : "int32",
              "default" : 20
            }
          }
        ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/AdminBookingView"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/booking-payments" : {
      "get" : {
        "tags" : [
          "admin-booking-payment-controller"
        ],
        "operationId" : "list_11",
        "parameters" : [
          {
            "name" : "status",
            "in" : "query",
            "required" : false,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "name" : "page",
            "in" : "query",
            "required" : false,
            "schema" : {
              "type" : "integer",
              "format" : "int32",
              "default" : 0
            }
          },
          {
            "name" : "size",
            "in" : "query",
            "required" : false,
            "schema" : {
              "type" : "integer",
              "format" : "int32",
              "default" : 20
            }
          }
        ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/AdminPaymentDeclarationView"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/audit" : {
      "get" : {
        "tags" : [
          "admin-audit-controller"
        ],
        "operationId" : "list_12",
        "parameters" : [
          {
            "name" : "action",
            "in" : "query",
            "required" : false,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "name" : "page",
            "in" : "query",
            "required" : false,
            "schema" : {
              "type" : "integer",
              "format" : "int32",
              "default" : 0
            }
          },
          {
            "name" : "size",
            "in" : "query",
            "required" : false,
            "schema" : {
              "type" : "integer",
              "format" : "int32",
              "default" : 50
            }
          }
        ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuditPage"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "TransferTicketRequest" : {
        "type" : "object",
        "properties" : {
          "firstName" : {
            "type" : "string",
            "maxLength" : 100,
            "minLength" : 0
          },
          "lastName" : {
            "type" : "string",
            "maxLength" : 100,
            "minLength" : 0
          },
          "email" : {
            "type" : [
              "string",
              "null"
            ],
            "format" : "email",
            "maxLength" : 255,
            "minLength" : 0
          },
          "phoneNumber" : {
            "type" : [
              "string",
              "null"
            ],
            "maxLength" : 32,
            "minLength" : 0
          }
        },
        "required" : [
          "firstName",
          "lastName"
        ]
      },
      "RsvpView" : {
        "type" : "object",
        "properties" : {
          "eventName" : {
            "type" : "string"
          },
          "eventWhen" : {
            "type" : [
              "string",
              "null"
            ]
          },
          "eventLocation" : {
            "type" : [
              "string",
              "null"
            ]
          },
          "organizerName" : {
            "type" : "string"
          },
          "primaryColor" : {
            "type" : "string"
          },
          "logoUrl" : {
            "type" : [
              "string",
              "null"
            ]
          },
          "guestName" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "ticketCode" : {
            "type" : [
              "string",
              "null"
            ]
          },
          "eventStatus" : {
            "type" : [
              "string",
              "null"
            ]
          },
          "erased" : {
            "type" : "boolean"
          },
          "transferAllowed" : {
            "type" : "boolean"
          },
          "transferDeadline" : {
            "type" : [
              "string",
              "null"
            ],
            "format" : "date-time"
          },
          "transferredTo" : {
            "type" : [
              "string",
              "null"
            ]
          }
        }
      },
...
      "ManualPaymentRequest" : {
        "type" : "object",
        "properties" : {
          "tier" : {
            "type" : "string",
            "minLength" : 1
          }
        },
        "required" : [
          "tier"
        ]
      },
      "ManualPaymentInstructions" : {
        "type" : "object",
        "properties" : {
          "paymentId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "reference" : {
            "type" : "string"
          },
          "tier" : {
            "type" : "string"
          },
          "amountMinor" : {
            "type" : "integer",
            "format" : "int64"
          },
          "currency" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "payee" : {
            "$ref" : "#/components/schemas/PayeeDetails"
          }
        }
      },
      "PayeeDetails" : {
        "type" : "object",
        "properties" : {
          "payeeName" : {
            "type" : [
              "string",
              "null"
            ]
          },
          "mobileMoneyNumber" : {
            "type" : [
              "string",
              "null"
            ]
          },
          "mobileMoneyOperator" : {
            "type" : [
              "string",
              "null"
            ]
          },
          "bankDetails" : {
            "type" : [
              "string",
              "null"
            ]
          }
        }
      },
...
      "RejectPaymentRequest" : {
        "type" : "object",
        "properties" : {
          "reason" : {
            "type" : "string",
            "minLength" : 1
          }
        },
        "required" : [
          "reason"
        ]
      },
      "AdminPaymentView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "tenantId" : {
            "type" : "string"
          },
          "eventId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "tier" : {
            "type" : "string"
          },
          "amountMinor" : {
            "type" : "integer",
            "format" : "int64"
          },
          "currency" : {
            "type" : "string"
          },
          "provider" : {
            "type" : "string"
          },
          "reference" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "ConfirmPaymentRequest" : {
        "type" : "object",
        "properties" : {
          "transactionReference" : {
            "type" : "string",
            "minLength" : 1
          }
        },
        "required" : [
          "transactionReference"
        ]
      },
      "AdminBookingCancellationView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "status" : {
            "type" : "string"
          },
          "refundAmountMinor" : {
            "type" : "integer",
            "format" : "int64"
          },
          "currency" : {
            "type" : "string"
          }
        }
      },
      "AdminPaymentDeclarationView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "bookingId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "customerName" : {
            "type" : "string"
          },
          "amountMinor" : {
            "type" : "integer",
            "format" : "int64"
          },
          "currency" : {
            "type" : "string"
          },
          "kind" : {
            "type" : "string"
          },
          "operator" : {
            "type" : "string"
          },
          "transactionReference" : {
            "type" : "string"
          },
          "declaredAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "verificationStatus" : {
            "type" : "string"
          },
          "verifiedBy" : {
            "type" : [
              "string",
              "null"
            ]
          },
          "verifiedAt" : {
            "type" : [
              "string",
              "null"
            ],
            "format" : "date-time"
          },
          "rejectionReason" : {
            "type" : [
              "string",
              "null"
            ]
          }
        }
      },
      "RejectDeclarationRequest" : {
        "type" : "object",
        "properties" : {
          "reason" : {
            "type" : "string",
            "minLength" : 1
          }
        },
        "required" : [
          "reason"
        ]
      },
...
      "ValidatorContextResponse" : {
        "type" : "object",
        "properties" : {
          "eventName" : {
            "type" : "string"
          },
          "eventStatus" : {
            "type" : "string"
          },
          "startDateTime" : {
            "type" : [
              "string",
              "null"
            ],
            "format" : "date-time"
          },
          "timezone" : {
            "type" : "string"
          },
          "eventLocation" : {
            "type" : [
              "string",
              "null"
            ]
          },
          "organizerName" : {
            "type" : "string"
          },
          "primaryColor" : {
            "type" : "string"
          },
          "logoUrl" : {
            "type" : [
              "string",
              "null"
            ]
          },
          "validatorLabel" : {
            "type" : "string"
          },
          "checkedIn" : {
            "type" : "integer",
            "format" : "int64"
          },
          "confirmed" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
      "AdminBookingView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "customerName" : {
            "type" : "string"
          },
          "customerPhone" : {
            "type" : "string"
          },
          "customerEmail" : {
            "type" : "string"
          },
          "eventType" : {
            "type" : "string"
          },
          "eventDate" : {
            "type" : "string",
            "format" : "date"
          },
          "guestCountEstimate" : {
            "type" : "integer",
            "format" : "int64"
          },
          "tier" : {
            "type" : "string"
          },
          "currency" : {
            "type" : "string"
          },
          "totalAmountMinor" : {
            "type" : "integer",
            "format" : "int64"
          },
          "depositAmountMinor" : {
            "type" : "integer",
            "format" : "int64"
          },
          "balanceAmountMinor" : {
            "type" : "integer",
            "format" : "int64"
          },
          "balanceDueDate" : {
            "type" : "string",
            "format" : "date"
          },
          "status" : {
            "type" : "string"
          },
          "tenantId" : {
            "type" : [
              "string",
              "null"
            ]
          },
          "eventId" : {
            "type" : [
              "string",
              "null"
            ],
            "format" : "uuid"
          },
          "acquisitionSource" : {
            "type" : [
              "string",
              "null"
            ]
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "AuditEntry" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "adminId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "action" : {
            "type" : "string"
          },
          "target" : {
            "type" : "string"
          },
          "note" : {
            "type" : [
              "string",
              "null"
            ]
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "AuditPage" : {
        "type" : "object",
        "properties" : {
          "entries" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AuditEntry"
            }
          },
          "total" : {
            "type" : "integer",
            "format" : "int64"
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      }
    },
    "securitySchemes" : {
      "bearer-jwt" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}